+++ src/dates.js
@@ -1,3 +1,3 @@
 export { default as formatDate } from './dates/format-date';
 export { default as parseDate } from './dates/parse-date';
-export { dateFormatString as dateFormatString } from './dates/format-date';
+export { default as dateFormatString } from './dates/date-format-string';
