+++ src/dates/date-format-string.js
@@ -1,0 +1,5 @@
+import dateGenericFormat from './date-generic-format';
+
+export default function dateFormatString(date, format, locale = "en") {
+    return dateGenericFormat(date, format, locale, true);
+}
